markup5ever: Move BufferQueue tests into that module.
